+++ files/NacoWiki/CHANGES.md
@@ -7,12 +7,15 @@ author: alex
 
 # 3.4.0-dev1
 
+- Added `vacuum` option.  If enabled `walkTree` (usually from searching
+  files) will delete metadata files that are orphaned.
 - Core: properties are now saved as JSON files (the PHP JSON parser is more reliable than the YAML one).
 - Util: added `is_program_in_path`
 - plugins:
   - Versions
     - Deta files are now saved as JSON files (the PHP JSON parser is more reliable than the YAML one).
     - Only gets enabled if `diff` and `patch` commands are available.
+    - Minor UI tweaks
 
 # 3.3.0-rel